修复了'get_civitai_info_from_hash'的BUG 迁移了api代码
